[Refactor] 영화 뷰 SRP 적용 (#9)
