+++ Tving_CloneProject/Tving_CloneProject/Scene/Main/Views/MovieView.swift
@@ -14,15 +14,10 @@ final class MovieView: UIView {
     
     // MARK: - UI Properties
     
-    private let dailyBoxOfficeCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
+    let dailyBoxOfficeCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
     
-    private var loadingIndicator = UIActivityIndicatorView()
-    
-    
-    // MARK: - Properties
-    
-    private var movieViewModel: MovieViewModel = MovieViewModel()
-    
+    let loadingIndicator = UIActivityIndicatorView()
+
     
     // MARK: - Life Cycles
     
@@ -32,10 +27,6 @@ final class MovieView: UIView {
         setHierarchy()
         setLayout()
         setStyle()
-        setDelegate()
-        registerCell()
-        setViewModel()
-        getDailyBoxOffice()
     }
     
     required init?(coder: NSCoder) {
@@ -69,74 +60,5 @@ private extension MovieView {
             $0.backgroundColor = UIColor(resource: .black)
         }
     }
-    
-    func setViewModel() {
-        movieViewModel.didUpdateNetworkResult.bind { [weak self] isSuccess in
-            guard let isSuccess else { return }
-            if isSuccess {
-                self?.dailyBoxOfficeCollectionView.reloadData()
-            }
-        }
-        
-        movieViewModel.didChangeLoadingIndicator.bind { [weak self] isLoading in
-            guard let isLoading else { return }
-            if isLoading {
-                self?.loadingIndicator.startAnimating()
-            } else {
-                self?.loadingIndicator.stopAnimating()
-            }
-        }
 
-    }
-    
-    func getDailyBoxOffice() {
-        if movieViewModel.getDailyBoxOffice() {
-            self.dailyBoxOfficeCollectionView.reloadData()
-        }
-    }
-    
-    func registerCell() {
-        dailyBoxOfficeCollectionView.register(DailyBoxOfficeCell.self, forCellWithReuseIdentifier: DailyBoxOfficeCell.identifier)
-    }
-    
-    func setDelegate() {
-        dailyBoxOfficeCollectionView.delegate = self
-        dailyBoxOfficeCollectionView.dataSource = self
-    }
-
-}
-
-extension MovieView: UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: Constant.Screen.width / 2 - 20, height: Constant.Screen.height / 3.5)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return 20
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return 10
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        return UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-    }
-    
-}
-
-extension MovieView: UICollectionViewDataSource {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return movieViewModel.fetchData().count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: DailyBoxOfficeCell.identifier, for: indexPath) as? DailyBoxOfficeCell else { return UICollectionViewCell() }
-        cell.setCell(contents: movieViewModel.fetchData()[indexPath.row])
-        
-        return cell
-    }
-    
 }
